Extract post processor logic out of document builder
