+++ src/Microsoft.DocAsCode.Build.Engine/DocumentBuilder.cs
@@ -13,13 +13,11 @@ namespace Microsoft.DocAsCode.Build.Engine
     using System.Reflection;
     using System.Text;
 
-    using Microsoft.DocAsCode.Build.Common;
     using Microsoft.DocAsCode.Build.Engine.Incrementals;
     using Microsoft.DocAsCode.Common;
     using Microsoft.DocAsCode.Dfm.MarkdownValidators;
     using Microsoft.DocAsCode.Exceptions;
     using Microsoft.DocAsCode.Plugins;
-
 
     public class DocumentBuilder : IDisposable
     {
@@ -30,7 +28,6 @@ namespace Microsoft.DocAsCode.Build.Engine
         internal IEnumerable<IInputMetadataValidator> MetadataValidators { get; set; }
 
         private readonly string _intermediateFolder;
-        private readonly List<PostProcessor> _postProcessors = new List<PostProcessor>();
         private readonly CompositionHost _container;
         private readonly BuildInfo _currentBuildInfo =
             new BuildInfo
@@ -38,7 +35,8 @@ namespace Microsoft.DocAsCode.Build.Engine
                 BuildStartTime = DateTime.UtcNow,
                 DocfxVersion = typeof(DocumentBuilder).Assembly.GetCustomAttribute<AssemblyFileVersionAttribute>()?.Version
             };
-        private BuildInfo _lastBuildInfo;
+        private readonly BuildInfo _lastBuildInfo;
+        private readonly HandlePostProcessors _handlePostProcessors;
 
         public DocumentBuilder(
             IEnumerable<Assembly> assemblies,
@@ -51,8 +49,9 @@ namespace Microsoft.DocAsCode.Build.Engine
             Logger.LogVerbose("Loading plug-in...");
             using (new LoggerPhaseScope("ImportPlugins", true))
             {
-                var assemblyList = assemblies?.ToList();
-                _container = GetContainer(assemblyList);
+                var assemblyList = assemblies?.ToList() ?? new List<Assembly>();
+                assemblyList.Add(typeof(DocumentBuilder).Assembly);
+                _container = CompositionUtility.GetContainer(assemblyList);
                 _container.SatisfyImports(this);
                 _currentBuildInfo.CommitFromSHA = commitFromSHA;
                 _currentBuildInfo.CommitToSHA = commitToSHA;
@@ -68,7 +67,6 @@ namespace Microsoft.DocAsCode.Build.Engine
             {
                 Logger.LogVerbose($"\t{processor.Name} with build steps ({string.Join(", ", from bs in processor.BuildSteps orderby bs.BuildOrder select bs.Name)})");
             }
-            _postProcessors = GetPostProcessor(postProcessorNames);
             if (intermediateFolder != null)
             {
                 var expanded = Environment.ExpandEnvironmentVariables(intermediateFolder);
@@ -79,6 +77,7 @@ namespace Microsoft.DocAsCode.Build.Engine
                 _intermediateFolder = Path.GetFullPath(expanded);
             }
             _lastBuildInfo = BuildInfo.Load(_intermediateFolder);
+            _handlePostProcessors = new HandlePostProcessors(_container, postProcessorNames);
         }
 
         public void Build(DocumentBuildParameters parameter)
@@ -97,7 +96,7 @@ namespace Microsoft.DocAsCode.Build.Engine
                 throw new ArgumentException("Parameters are empty.", nameof(parameters));
             }
 
-            var markdownServiceProvider = GetExport<IMarkdownServiceProvider>(parameters[0].MarkdownEngineName);
+            var markdownServiceProvider = CompositionUtility.GetExport<IMarkdownServiceProvider>(_container, parameters[0].MarkdownEngineName);
             if (markdownServiceProvider == null)
             {
                 Logger.LogError($"Unable to find markdown engine: {parameters[0].MarkdownEngineName}");
@@ -121,7 +120,7 @@ namespace Microsoft.DocAsCode.Build.Engine
                 {
                     transformDocument = true;
                 }
-                PrepareMetadata(parameter);
+                _handlePostProcessors.PrepareMetadata(parameter.Metadata);
                 if (!string.IsNullOrEmpty(parameter.VersionName))
                 {
                     Logger.LogInfo($"Start building for version: {parameter.VersionName}");
@@ -138,10 +137,10 @@ namespace Microsoft.DocAsCode.Build.Engine
                 .ReadFromRealFileSystem(parameters[0].OutputBaseDir)
                 .WriteToRealFileSystem(parameters[0].OutputBaseDir)
                 .Create();
-            var generatedManifest = MergeManifest(manifests);
+            var generatedManifest = ManifestUtility.MergeManifest(manifests);
 
-            RemoveDuplicateOutputFiles(generatedManifest.Files);
-            PostProcess(generatedManifest, outputDirectory);
+            ManifestUtility.RemoveDuplicateOutputFiles(generatedManifest.Files);
+            _handlePostProcessors.PostProcess(generatedManifest, outputDirectory);
 
             // Save to manifest.json
             SaveManifest(generatedManifest);
@@ -184,181 +183,15 @@ namespace Microsoft.DocAsCode.Build.Engine
             }
             catch (Exception ex)
             {
-                Logger.LogWarning($"Fail to init markdown style, details:{Environment.NewLine}{ex.ToString()}");
+                Logger.LogWarning($"Fail to init markdown style, details:{Environment.NewLine}{ex.Message}");
                 return Enumerable.Empty<IInputMetadataValidator>();
             }
-        }
-
-        private void PrepareMetadata(DocumentBuildParameters parameters)
-        {
-            foreach (var postProcessor in _postProcessors)
-            {
-                using (new LoggerPhaseScope($"Prepare metadata in post processor {postProcessor.ContractName}", false))
-                using (new PerformanceScope($"Prepare metadata in post processor {postProcessor.ContractName}", LogLevel.Verbose))
-                {
-                    parameters.Metadata = postProcessor.Processor.PrepareMetadata(parameters.Metadata);
-                    if (parameters.Metadata == null)
-                    {
-                        throw new DocfxException($"Plugin {postProcessor.ContractName} should not return null metadata");
-                    }
-                }
-            }
-        }
-
-        private void PostProcess(Manifest manifest, string outputDir)
-        {
-            // post process
-            foreach (var postProcessor in _postProcessors)
-            {
-                using (new LoggerPhaseScope($"Process in post processor {postProcessor.ContractName}", false))
-                using (new PerformanceScope($"Process in post processor {postProcessor.ContractName}", LogLevel.Verbose))
-                {
-                    manifest = postProcessor.Processor.Process(manifest, outputDir);
-                    if (manifest == null)
-                    {
-                        throw new DocfxException($"Plugin {postProcessor.ContractName} should not return null manifest");
-                    }
-
-                    // To make sure post processor won't generate duplicate output files
-                    RemoveDuplicateOutputFiles(manifest.Files);
-                }
-            }
-        }
-
-        private void RemoveDuplicateOutputFiles(List<ManifestItem> manifestItems)
-        {
-            if (manifestItems == null)
-            {
-                throw new ArgumentNullException(nameof(manifestItems));
-            }
-
-            var itemsToRemove = new HashSet<string>();
-            foreach (var duplicates in (from m in manifestItems
-                                        from output in m.OutputFiles.Values
-                                        let relativePath = output?.RelativePath
-                                        select new { item = m, relativePath = relativePath })
-                              .GroupBy(obj => obj.relativePath, FilePathComparer.OSPlatformSensitiveStringComparer)
-                              .Where(g => g.Count() > 1))
-            {
-                Logger.LogWarning($"Overwrite occurs while input files \"{string.Join(", ", duplicates.Select(duplicate => duplicate.item.SourceRelativePath))}\" writing to the same output file \"{duplicates.Key}\"");
-                itemsToRemove.UnionWith(duplicates.Skip(1).Select(duplicate => duplicate.item.SourceRelativePath));
-            }
-            manifestItems.RemoveAll(m => itemsToRemove.Contains(m.SourceRelativePath));
-        }
-
-        private static Manifest MergeManifest(List<Manifest> manifests)
-        {
-            var xrefMaps = (from manifest in manifests
-                            where manifest.XRefMap != null
-                            select manifest.XRefMap).ToList();
-            var incrementalInfos = (from manifest in manifests
-                                    from i in manifest.IncrementalInfo ?? Enumerable.Empty<IncrementalInfo>()
-                                    select i).ToList();
-            return new Manifest
-            {
-                Homepages = (from manifest in manifests
-                             from homepage in manifest.Homepages ?? Enumerable.Empty<HomepageInfo>()
-                             select homepage).Distinct().ToList(),
-                Files = (from manifest in manifests
-                         from file in manifest.Files ?? Enumerable.Empty<ManifestItem>()
-                         select file).Distinct().ToList(),
-                XRefMap = xrefMaps.Count <= 1 ? xrefMaps.FirstOrDefault() : xrefMaps,
-                SourceBasePath = manifests.FirstOrDefault()?.SourceBasePath,
-                IncrementalInfo = incrementalInfos.Count > 0 ? incrementalInfos : null,
-            };
         }
 
         private static void SaveManifest(Manifest manifest)
         {
             JsonUtility.Serialize(Constants.ManifestFileName, manifest);
             Logger.LogInfo($"Manifest file saved to {Constants.ManifestFileName}.");
-        }
-
-        private List<PostProcessor> GetPostProcessor(ImmutableArray<string> processors)
-        {
-            var processorList = new List<PostProcessor>();
-            AddBuildInPostProcessor(processorList);
-            foreach (var processor in processors)
-            {
-                var p = GetExport<IPostProcessor>(processor);
-                if (p != null)
-                {
-                    processorList.Add(new PostProcessor
-                    {
-                        ContractName = processor,
-                        Processor = p
-                    });
-                    Logger.LogInfo($"Post processor {processor} loaded.");
-                }
-                else
-                {
-                    Logger.LogWarning($"Can't find the post processor: {processor}");
-                }
-            }
-            return processorList;
-        }
-
-        private static void AddBuildInPostProcessor(List<PostProcessor> processorList)
-        {
-            processorList.Add(
-                new PostProcessor
-                {
-                    ContractName = "html",
-                    Processor = new HtmlPostProcessor
-                    {
-                        Handlers =
-                        {
-                            new ValidateBookmark(),
-                            new RemoveDebugInfo(),
-                        },
-                    }
-                });
-        }
-
-        private T GetExport<T>(string name) where T : class =>
-            (T)GetExport(typeof(T), name);
-
-        private object GetExport(Type type, string name)
-        {
-            object exportedObject = null;
-            try
-            {
-                exportedObject = _container.GetExport(type, name);
-            }
-            catch (CompositionFailedException ex)
-            {
-                Logger.LogWarning($"Can't import: {name}, {ex}");
-            }
-            return exportedObject;
-        }
-
-        private static CompositionHost GetContainer(IEnumerable<Assembly> assemblies)
-        {
-            var configuration = new ContainerConfiguration();
-
-            configuration.WithAssembly(typeof(DocumentBuilder).Assembly);
-
-            if (assemblies != null)
-            {
-                foreach (var assembly in assemblies)
-                {
-                    if (assembly != null)
-                    {
-                        configuration.WithAssembly(assembly);
-                    }
-                }
-            }
-
-            try
-            {
-                return configuration.CreateContainer();
-            }
-            catch (ReflectionTypeLoadException ex)
-            {
-                Logger.LogError(
-                    $"Error when get composition container: {ex.Message}, loader exceptions: {(ex.LoaderExceptions != null ? string.Join(", ", ex.LoaderExceptions.Select(e => e.Message)) : "none")}");
-                throw;
-            }
         }
 
         private static string ComputePluginHash(List<Assembly> assemblyList)
@@ -375,24 +208,14 @@ namespace Microsoft.DocAsCode.Build.Engine
                 {
                     builder.AppendLine(item);
                 }
-                return StringExtension.GetMd5String(builder.ToString());
+                return builder.ToString().GetMd5String();
             }
             return string.Empty;
         }
 
         public void Dispose()
         {
-            foreach (var processor in _postProcessors)
-            {
-                Logger.LogVerbose($"Disposing processor {processor.ContractName} ...");
-                (processor.Processor as IDisposable)?.Dispose();
-            }
-        }
-
-        private sealed class PostProcessor
-        {
-            public string ContractName { get; set; }
-            public IPostProcessor Processor { get; set; }
+            _handlePostProcessors.Dispose();
         }
     }
 }
