+++ src/Microsoft.DocAsCode.Common/project.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "Microsoft.DocAsCode.Plugins": "1.0.0",
     "Microsoft.DocAsCode.YamlSerialization": "1.0.0",
+    "Microsoft.Composition": "1.0.27",
     "YamlDotNet.Signed": "3.9.0",
     "Newtonsoft.Json": "9.0.1",
     "System.Collections.Immutable": "1.2.0"
